StewardBot: monkey patch for missing execute_every method
In 8f9f85e, a call to `self.execute_every` was introduced to clean stale
entries out of the attention use cache. This method is not present in
the SingleServerIRCBot base class. StewardBot currently relies on
a globally installed irclib package. On the Tool Labs Trusty exec nodes,
the provided version is irclib 0.4.8. The execute_every method was not
added to irclib until 0.6.0.

Add a monkey patch method with equivalent functionality until we can get
things updated by installing a modern irclib version in a virtualenv.

Change-Id: Iaba3ad8a973dd7eed73e3311dcdba6656bdeaab3