Keep dynamic list of Stewards
Add the very basic function of fetching all accounts with the Steward
flag on the account and keeping it in memory within a list. This will
be the very basic foundations of monitoring Steward accounts for being
locked which may indicate an emergent situation involving a compromised
account. This commit will add a unremarkable !steward ping as a response

Change-Id: I05b0f68e1ba3a4f92f09a76e4f9a428252303705