SULWatcher: Use target username instead of acting user
Currently, SULWatcher reports change["user"] when an account is created.
For `create` entries, this is fine, as it matches the created username.
For `create2` and `byemail` creations, however, this is the name of the
acting user and not the created username. This increases confusion, and
makes it harder to tell at a glance if an account is impersonating
someone.

With this patch, SULWatcher will always report the created username. The
blacklist is always applied to the created username, and the whitelist
is always applied to the creating user. This means creations by
whitelisted users are still not displayed.

Change-Id: I598c42ff4d804b87abe2d1f8f21a8501ecc09240