Delete privacy.html
Dates back for when our webservice used lighttpd and access.log
was enabled by default,but this is is no longer our the case.

The current uwsgi.log does not seem to be collecting any PII
(IP is always a private one) and stewardbots is not a project
that lets others create accounts so no private information is
being collected. AFAIK we are not using cookies either but if we
do a footer to the WMF cookie statement would probably
be all we need.

In addition, stewardbots.toolforge uses the stewardbots-ui repo
which already has proper footers to the WMCS ToU and the TCoC.

Bug: T130031
Change-Id: I0c3c467e749995c3eff9baffe8257000e2e48ef6