StewardBot: add health check script based on canary events
This changeset adds a health check script (k8s liveliness probe) to
automatically restart the StewardBot pod if the RC reader fails.
https://wikitech.wikimedia.org/wiki/Help:Toolforge/Jobs_framework#Configuring_health-checks_for_jobs

If configured to do so, StewardBot will touch a /tmp file every time it
recieves a canary event. The health check script will then check if the
/tmp file exists and has not been modified for 65 minutes, restarting
the pod if so. If the /tmp file does not exist or has been touched
recently, nothing happens.

Focusing only on the hourly canary events will prevent unnecessary
restarts due to outages, but limits how quickly the bot will restart
after failures. If this proves to still be too slow, the conditions that
run heartbeat() can be changed in the future.

Change-Id: Ib2b36d0ce88a1a343d878f5d529de1505f1a7707